+++ clang-tools-extra/clang-tidy/misc/MiscTidyModule.cpp
@@ -13,6 +13,7 @@
 #include "ConstCorrectnessCheck.h"
 #include "CoroutineHostileRAIICheck.h"
 #include "DefinitionsInHeadersCheck.h"
+#include "DoubleSharedPtrCheck.h"
 #include "ExplicitConstructorCheck.h"
 #include "HeaderIncludeCycleCheck.h"
 #include "IncludeCleanerCheck.h"
@@ -55,6 +56,8 @@ public:
         "misc-coroutine-hostile-raii");
     CheckFactories.registerCheck<DefinitionsInHeadersCheck>(
         "misc-definitions-in-headers");
+    CheckFactories.registerCheck<DoubleSharedPtrCheck>(
+        "misc-double-shared-ptr");
     CheckFactories.registerCheck<ExplicitConstructorCheck>(
         "misc-explicit-constructor");
     CheckFactories.registerCheck<HeaderIncludeCycleCheck>(